Move EXT0 and UART
